requirements.txt
-i https://pypi.org/simple
distro==1.9.0; python_version >= '3.6'
jpype1==1.5.0
numpy==1.26.3; python_version < '3.11'
packaging==23.2; python_version >= '3.7'
pandas==2.1.4; python_version >= '3.9'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3.post1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tabula-py[jpype]==2.9.0; python_version >= '3.8'
tzdata==2023.4; python_version >= '2'
